+++ jsonargparse/_loaders_dumpers.py
@@ -126,8 +126,8 @@ def jsonnet_load(stream, path="", ext_vars=None):
     _jsonnet = import_jsonnet("jsonnet_load")
     path_dir = current_local_dir.get()
     if path_dir and not os.path.isabs(path):
-        # jsonnet resolves imports relative to the given file name
-        path = os.path.join(path_dir, path or "snippet")
+        # jsonnet resolves imports relative to the given file name, which path_dir already accounts for
+        path = os.path.join(path_dir, os.path.basename(path) or "snippet")
     try:
         val = _jsonnet.evaluate_snippet(path, stream, ext_vars=ext_vars, ext_codes=ext_codes)
     except RuntimeError: